backend/requirements.txt
tensorflow>=2.18.0
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow>=10.4.0
numpy>=2.1.0
scikit-image>=0.24.0
python-dotenv==1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0